fix: RadarChart crashes if all values in RadarDataSet are equal
If the max and min values in a RadarDataSet are equal, the `drawTicks` method in `radar_chart_painter.dart` will enter an infinite loop.

This is due to a for loop that fails to be incremented if the max and min values are equal, causing the loop to never meet its exit condition.

I have replaced the for loop with one that will always exit.

Resolves #882